techpack/audio: Mark availability of FSA4490 i2c node as unlikely
Our device doesn't i2c for Fairchild FSA4480 IC switch chip.
So, the conditional check to check for the availability of the
required node is definitely going to fail as we don't have that
node in our proprietary vendor device tree. So, mark the first
condition as unlikely as it's never going to happen. Compiler
can take this hint and proceed to mark fsa_enable as false.
While at it, silence the log for the unavailability of FSA4480
i2c node. It's an inappropriate use of dev_err() since this is
not even an error but just our device not supporting the node.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: pix106 <sbordenave@gmail.com>